feature: Added header, footer, and home button to the book reviews page. Added the webfolio as another project in the projects section
