+++ app/book-reviews/reviews-footer.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default function ReviewsFooter() {
+  return (
+    <footer className = "align-middle px-4 text-center text-gray-500">
+        <p className = "text-xs">
+            <span className="font-semibold">About this website: </span>
+            Built with React/NextJS, TypeScript, and Tailwind CSS
+        </p>
+    </footer>
+  )
+}
